backend/python/requirements.txt
flask==2.3.3
flask-cors==4.0.0
numpy==1.24.3
opencv-python==4.8.0.76
scikit-image==0.21.0
scikit-learn==1.3.0
torch==2.0.1
torchvision==0.15.2
sqlalchemy==2.0.20
alembic==1.12.0
python-dotenv==1.0.0
pytest==7.4.0
pillow==10.0.0
matplotlib==3.7.2
pandas==2.0.3
tqdm==4.65.0 